BiDi: consolidate BiDi fixtures in webdriver/tests/support/fixtures_bidi
This fixtures file is already globally included in
webdriver/tests/conftest.py via `pytest_plugins`.

Stop having webdriver/tests/bidi/conftest.py be a duplicate of the
support fixture.
