--- webdriver/tests/bidi/conftest.py
@@ -1,48 +1,0 @@
-import pytest
-
-
-@pytest.fixture
-def test_origin(url):
-    return url("")
-
-
-@pytest.fixture
-def test_alt_origin(url):
-    return url("", domain="alt")
-
-
-@pytest.fixture
-def test_page(inline):
-    return inline("<div>foo</div>")
-
-
-@pytest.fixture
-def test_page2(inline):
-    return inline("<div>bar</div>")
-
-
-@pytest.fixture
-def test_page_cross_origin(inline):
-    return inline("<div>bar</div>", domain="alt")
-
-
-@pytest.fixture
-def test_page_multiple_frames(inline, test_page, test_page2):
-    return inline(
-        f"<iframe src='{test_page}'></iframe><iframe src='{test_page2}'></iframe>"
-    )
-
-
-@pytest.fixture
-def test_page_nested_frames(inline, test_page_same_origin_frame):
-    return inline(f"<iframe src='{test_page_same_origin_frame}'></iframe>")
-
-
-@pytest.fixture
-def test_page_cross_origin_frame(inline, test_page_cross_origin):
-    return inline(f"<iframe src='{test_page_cross_origin}'></iframe>")
-
-
-@pytest.fixture
-def test_page_same_origin_frame(inline, test_page):
-    return inline(f"<iframe src='{test_page}'></iframe>")
